02-memorygpt-rag-chatbot/requirements.txt
streamlit
python-dotenv
langchain
chromadb
sentence-transformers
pypdf
